arm: dts: orangepi: Fix Orange Pi R1 and Plus LTS boot issue
Orange Pi R1 Plus and LTS don't have eMMC (default mmc0 in rk3328-u-boot.dtsi).
The device is equipped with SPI flash memory, which U-Boot tries to use for
reading configuration. However, the manufacturer has not provided any means
to utilize SPI flash memory. There is no documentation or software available
for this purpose.

As a result, during OpenWRT sysupgrade, the board fails to boot because the
script that switches the boot_device to an available one does not trigger.

This patch changes the alias of mmc0 from &emmc to &sdmmc, ensuring the board
boots correctly and sysupgrade completes successfully. Additionally, this
allows saving the U-Boot environment using saveenv with uboot-envtool.

Signed-off-by: Pugemon <islavaivanov76@gmail.com>
Cc: Kever Yang <kever.yang@rock-chips.com>